% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.partition.R
\name{b1part-ergmConstraint}
\alias{b1part-ergmConstraint}
\alias{InitErgmConstraint.b1part}
\title{Sample over partitioning of Mode 1 nodes identified by edges to Mode 2 nodes}
\usage{
# b1part
}
\description{
For bipartite networks, preserve the degree for the
first mode of each vertex at 1, while the second mode is used to
define a partitioning, with an appropriate combinatoric adjustment, if necessary.
}
\seealso{
\code{\link{ergmConstraint}} for index of constraints and hints currently visible to the package.

\Sexpr[results = rd, stage = render]{ergm:::.formatTermKeywords("ergmConstraint", "b1part", "subsection")}
}
\concept{bipartite}
