% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmProposal.partition.R
\name{B1Part-ergmProposal}
\alias{B1Part-ergmProposal}
\alias{InitErgmProposal.B1Part}
\title{MHp for b1part constraints}
\description{
MHp for \code{constraints= ~b1part}. At this time, this is
just the \code{\link[=CondB1Degree-ergmProposal]{CondB1Degree}} proposal
with a combinatoric adjustment.
}
\details{
\if{html}{\Sexpr[results = rd, stage = render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal = "B1Part"))}}
\if{text}{\Sexpr[results = rd, stage = render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal = "B1Part"))}}
\if{latex}{\Sexpr[results = rd, stage = render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal = "B1Part"))}}
}
\seealso{
\code{\link{ergmProposal}} for index of proposals currently visible to the package.

\Sexpr[results = rd, stage = render]{ergm:::.formatTermKeywords("ergmProposal", "B1Part", "subsection")}
}
\keyword{internal}
