% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.b1partblockdiag.R
\name{B1PartBlockdiag-ergmProposal}
\alias{B1PartBlockdiag-ergmProposal}
\alias{InitErgmProposal.B1PartBlockdiag}
\title{MHp for b1partblockdiag constraints}
\description{
MHp for \code{constraints = ~b1partblockdiag}. Reuses the
\code{B1Part} C proposal; the \code{free_dyads} field from the constraint
restricts sampling to within-block dyads.
}
\details{
\if{html}{\Sexpr[results = rd, stage = render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal = "B1PartBlockdiag"))}}
\if{text}{\Sexpr[results = rd, stage = render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal = "B1PartBlockdiag"))}}
\if{latex}{\Sexpr[results = rd, stage = render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal = "B1PartBlockdiag"))}}
}
\seealso{
\code{\link{ergmProposal}} for index of proposals currently visible to the package.

\Sexpr[results = rd, stage = render]{ergm:::.formatTermKeywords("ergmProposal", "B1PartBlockdiag", "subsection")}
}
\keyword{internal}
