% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exactcalculations.R
\name{calculate_denominator_Dirichlet_restricted}
\alias{calculate_denominator_Dirichlet_restricted}
\title{Estimation basic model with size constraint}
\usage{
calculate_denominator_Dirichlet_restricted(n, smin, smax, alpha, results)
}
\arguments{
\item{n}{XXX}

\item{smin}{XXX}

\item{smax}{XXX}

\item{alpha}{XXX}

\item{results}{XXX}
}
\value{
XXX
}
\description{
Estimation basic model with size constraint
}
