% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phase1.R
\name{calculate_inverted_covariance_and_scaling}
\alias{calculate_inverted_covariance_and_scaling}
\title{Calculation of the inverse of the covariance matrix and the scaling matrix}
\usage{
calculate_inverted_covariance_and_scaling(
  startingestimates,
  z.obs,
  nodes,
  effects,
  objects,
  a.scaling,
  length.phase,
  z.phase,
  fixed.estimates
)
}
\arguments{
\item{startingestimates}{XXX}

\item{z.obs}{XXX}

\item{nodes}{node set (data frame)}

\item{effects}{effects/sufficient statistics (list with a vector "names", and a vector "objects")}

\item{objects}{objects used for statistics calculation (list with a vector "name", and a vector "object")}

\item{a.scaling}{XXX}

\item{length.phase}{XXX}

\item{fixed.estimates}{XXX}
}
\value{
XXX
}
\description{
Calculation of the inverse of the covariance matrix and the scaling matrix
}
