% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exactcalculations.R
\name{calculate_proba_Dirichlet_restricted}
\alias{calculate_proba_Dirichlet_restricted}
\title{Calculate Dirichlet probability}
\usage{
calculate_proba_Dirichlet_restricted(alpha, stat, n, smin, smax)
}
\arguments{
\item{alpha}{parameter value}

\item{stat}{observed stat (number of groups)}

\item{n}{number of nodes}

\item{smin}{minimum size for a group}

\item{smax}{maximum size for a group}
}
\value{
a numeric
}
\description{
Calculate the probability of observing a partition with a given number of groups
for a model with a single statistic for the number of groups and a given parameter value.
The set of possible partitions can be restricted to partitions with groups of a certain size.
}
