% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{check_sizes}
\alias{check_sizes}
\title{Function to determine whether a partition contains the allowed group sizes}
\usage{
check_sizes(partition, sizes.allowed)
}
\arguments{
\item{partition}{XXX}

\item{sizes.allowed}{XXX}
}
\value{
XXX
}
\description{
Function to determine whether a partition contains the allowed group sizes
}
