% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exactcalculations.R
\name{compute_averagesize}
\alias{compute_averagesize}
\title{Recursive function to compute the value of the denominator}
\usage{
compute_averagesize(num.nodes)
}
\arguments{
\item{num.nodes}{XXX}
}
\value{
XXX
}
\description{
Recursive function to compute the value of the denominator
}
