% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exactcalculations.R
\name{compute_numgroups_denominator}
\alias{compute_numgroups_denominator}
\title{Compute denominator for model with number of groups}
\usage{
compute_numgroups_denominator(num.nodes, alpha)
}
\arguments{
\item{num.nodes}{number of nodes}

\item{alpha}{parameter value}
}
\value{
a numeric
}
\description{
Recursive function to compute the value of the denominator for the model
with a single statistic which is the number of groups
}
