% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Metropolis.R
\name{compute_size_neighborhood_p1}
\alias{compute_size_neighborhood_p1}
\title{compute size neighborhood phase 1}
\usage{
compute_size_neighborhood_p1(partition)
}
\value{
A list
}
\description{
compute size neighborhood phase 1
}
