% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Metropolis.R
\name{compute_size_neighborhood_p1_restricted}
\alias{compute_size_neighborhood_p1_restricted}
\title{compute size neighborhood 1 restricted}
\usage{
compute_size_neighborhood_p1_restricted(partition, sizes.simulated)
}
\arguments{
\item{partition}{XXX}

\item{sizes.simulated}{XXX}
}
\value{
A list
}
\description{
compute size neighborhood 1 restricted
}
