% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Metropolis.R
\name{compute_size_neighborhood_p2}
\alias{compute_size_neighborhood_p2}
\title{compute size neighborhood 2}
\usage{
compute_size_neighborhood_p2(partition)
}
\arguments{
\item{partition}{XXX}
}
\value{
A list
}
\description{
compute size neighborhood 2
}
