% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Metropolis.R
\name{compute_size_neighborhood_p3_restricted}
\alias{compute_size_neighborhood_p3_restricted}
\title{compute size neighborhood 3 restricted}
\usage{
compute_size_neighborhood_p3_restricted(
  partition,
  numgroups.simulated,
  sizes.simulated
)
}
\arguments{
\item{partition}{XXX}

\item{numgroups.simulated}{XXX}

\item{sizes.simulated}{XXX}
}
\value{
A list
}
\description{
compute size neighborhood 3 restricted
}
