% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Metropolis.R
\name{compute_size_neighborhood_p4}
\alias{compute_size_neighborhood_p4}
\title{compute size neighborhood phase 4}
\usage{
compute_size_neighborhood_p4(partition)
}
\arguments{
\item{partition}{a partition}
}
\value{
A list
}
\description{
compute size neighborhood phase 4
}
