% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Metropolis.R
\name{compute_size_neighborhood_p7}
\alias{compute_size_neighborhood_p7}
\title{compute size neighborhood phase 7}
\usage{
compute_size_neighborhood_p7(partition, partition2)
}
\arguments{
\item{partition}{a partition}
}
\value{
A list
}
\description{
compute size neighborhood phase 7
}
