% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_change_statistics.R
\name{computeStatistics}
\alias{computeStatistics}
\title{Compute Statistics}
\usage{
computeStatistics(partition, nodes, effects, objects)
}
\arguments{
\item{partition}{vector, A partition}

\item{nodes}{data frame, Node set}

\item{effects}{list with a vector "names", and a vector "objects", Effects/sufficient statistics}

\item{objects}{list with a vector "name", and a vector "object", Objects used for statistics calculation}
}
\value{
the statistics
}
\description{
Compute Statistics
}
