% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_change_statistics.R
\name{computeStatistics_multiple}
\alias{computeStatistics_multiple}
\title{Compute Statistics multiple}
\usage{
computeStatistics_multiple(
  partitions,
  presence.tables,
  nodes,
  effects,
  objects,
  single.obs = NULL
)
}
\arguments{
\item{partitions}{Observed partitions}

\item{presence.tables}{to indicate which nodes were present when}

\item{nodes}{Node set (data frame)}

\item{effects}{Effects/sufficient statistics (list with a vector "names", and a vector "objects")}

\item{objects}{Objects used for statistics calculation (list with a vector "name", and a vector "object")}

\item{single.obs}{equal NULL by default}
}
\value{
A list
}
\description{
Function that computes the statistic vector for given (multiple) partitions and a given model
}
