% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_statistics.R
\name{correlation_with_size}
\alias{correlation_with_size}
\title{Correlation with size}
\usage{
correlation_with_size(partition, attribute, categorical)
}
\arguments{
\item{partition}{A partition (vector)}

\item{attribute}{A vector containing the values of the attribute}

\item{categorical}{A Boolean (True or False) indicating if the attribute is categorical}
}
\value{
A number corresponding to the correlation coefficient if the attribute is numerical or
the correlation ratio if the attribute is categorical.
}
\description{
This function computes the correlation between an attribute and the size of the groups.
}
\examples{
p <- c(1,2,2,3,3,4,4,4,5)
at <- c(3,5,23,2,1,0,3,9,2)
at2 <- c(3,5,20,2,1,0,0,9,0)
correlation_with_size(p,at,F)
}
