% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_utility.R
\name{correlation_within}
\alias{correlation_within}
\title{Within groups correlation}
\usage{
correlation_within(partition, attribute1, attribute2, group)
}
\arguments{
\item{partition}{A partition (vector)}

\item{attribute1}{A vector containing the values of the first attribute}

\item{attribute2}{A vector containing the values of the second attribute}

\item{group}{A number indicating the selected group}
}
\value{
A number corresponding to the correlation coefficient
}
\description{
This function computes the correlation between the two attributes for individuals in the same group.
}
\examples{
p <- c(1,2,2,3,3,4,4,4,5)
at <- c(3,5,23,2,1,0,3,9,2)
at2 <- c(3,5,20,2,1,0,0,9,0)
correlation_within(p,at,at2,4)
}
