% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_utility.R
\name{CUG}
\alias{CUG}
\title{CUG}
\usage{
CUG(observation, fun, permutations = NULL, num.permutations = 1000)
}
\arguments{
\item{observation}{The observed partition}

\item{fun}{An utility function or statistic to compute}

\item{permutations}{A predicted partition or partition to compare to observation. NULL by default}

\item{num.permutations}{Number of permutations to do to generate a partition if permutation = NULL. 1000 by default}
}
\value{
The value of mean, standard deviation, pvalue, and confidence interval of the estimator of the statistic in fun
}
\description{
This function computes ??
}
\examples{
p <- c(1,2,2,3,3,4,4,4,5)
CUG(p,fun=function(part){compute_number_same_pairs(part,at3,'avg')})
}
