% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_utility.R
\name{density}
\alias{density}
\title{Density of intra ties}
\usage{
density(partition, matrix, stat)
}
\arguments{
\item{partition}{A partition (vector)}

\item{matrix}{A matrix containing the values of the dyadic attribute}

\item{stat}{The statistic to compute : 'avg' for the average and 'sd' for the standard deviation}
}
\value{
The statisic chosen in stat.
}
\description{
This function computes the average "density" of dyadic attributes within groups and the
dispersion of the "density" of dyadic attributes within groups.
}
\examples{
p <- c(1,2,2,3,3,4,4,4,5)
mat <- matrix(1:81, ncol = 9)
density(p,mat,'avg')
}
