% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onUnload}
\alias{.onUnload}
\title{Package unload hook for ERPM}
\usage{
.onUnload(libpath)
}
\arguments{
\item{libpath}{Path to the library.}
}
\description{
Ensures that the ERPM shared library is properly unloaded when the
package is detached.
}
\keyword{internal}
