% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.RegisterProposals}
\alias{.RegisterProposals}
\title{Ensure legacy B1Part proposal row is present}
\usage{
.RegisterProposals()
}
\value{
Invisibly returns 'TRUE' if the row was added, 'FALSE' otherwise.
}
\description{
This function checks whether the row already exists
before attempting to add it again.
}
\details{
Note : The naming convention of this function follows the pre-existing code.
}
\keyword{internal}
