% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Metropolis.R
\name{draw_step_single_before}
\alias{draw_step_single_before}
\title{Draw step single before}
\usage{
draw_step_single_before(
  theta,
  current.partition,
  current.logit,
  current.z,
  nodes,
  effects,
  objects,
  neighborhood,
  sizes.allowed,
  sizes.simulated
)
}
\arguments{
\item{theta}{model parameters}

\item{nodes}{node set (data frame)}

\item{effects}{effects/sufficient statistics (list with a vector "names", and a vector "objects")}

\item{objects}{objects used for statistics calculation (list with a vector "name", and a vector "object")}

\item{neighborhood}{way of choosing partitions: probability vector (2 actors swap, merge/division, single actor move, single pair move, 2 pairs swap, 2 groups reshuffle)}

\item{sizes.allowed}{vector of group sizes allowed in sampling (now, it only works for vectors like size_min:size_max)}

\item{sizes.simulated}{vector of group sizes allowed in the Markov chain but not necessraily sampled (now, it only works for vectors like size_min:size_max)}
}
\value{
A list
}
\description{
Draw step single before
}
