% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erpm_wrapper.R
\name{erpm}
\alias{erpm}
\title{ERPM main wrapper: translate and optionally fit with \pkg{ergm}}
\usage{
erpm(
  formula,
  eval.call = TRUE,
  verbose = TRUE,
  estimate = NULL,
  eval.loglik = NULL,
  control = NULL,
  timeout = NULL,
  nodes = NULL,
  dyads = list(),
  group_labels = NULL
)
}
\description{
ERPM main wrapper: translate and optionally fit with \pkg{ergm}
}
