% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erpm_long_empile_helpers.R
\name{erpm_long_empile_helpers}
\alias{erpm_long_empile_helpers}
\title{ERPM longitudinal helpers (PLE / stacked) (internal)}
\description{
Placeholder for PLE-specific helpers.

IMPORTANT:
Not implemented in this refactor step.
}
\note{
erpm_long_empile_helpers.R
}
\keyword{ERGM}
\keyword{ERPM}
\keyword{helpers}
\keyword{internal}
\keyword{longitudinal}
