% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_estimate.R
\name{estimate_ERPM}
\alias{estimate_ERPM}
\title{Estimate ERPM}
\usage{
estimate_ERPM(
  partition,
  nodes,
  objects,
  effects,
  startingestimates,
  multiplicationfactor = 30,
  gainfactor = 0.1,
  a.scaling = 0.2,
  r.truncation.p1 = 2,
  r.truncation.p2 = 5,
  burnin = 30,
  thining = 10,
  length.p1 = 100,
  min.iter.p2 = NULL,
  max.iter.p2 = NULL,
  multiplication.iter.p2 = 100,
  num.steps.p2 = 6,
  length.p3 = 1000,
  neighborhood = c(0.7, 0.3, 0),
  fixed.estimates = NULL,
  sizes.allowed = NULL,
  sizes.simulated = NULL,
  double.averaging = F,
  inv.zcov = NULL,
  inv.scaling = NULL,
  parallel = F,
  parallel2 = F,
  cpus = 1
)
}
\arguments{
\item{partition}{observed partition}

\item{nodes}{nodeset (data frame)}

\item{objects}{objects used for statistics calculation (list with a vector "name", and a vector "object")}

\item{effects}{effects/sufficient statistics (list with a vector "names", and a vector "objects")}

\item{startingestimates}{first guess for the model parameters}

\item{multiplicationfactor}{= 30,  for now, useless}

\item{gainfactor}{= 0.1, numeric used to decrease the size of steps made in the Newton optimization}

\item{a.scaling}{= 0.2,  numeric used to reduce the influence of non-diagonal elements in the scaling matrix (for stability)}

\item{r.truncation.p1}{= 2 numeric used to limit extreme values in the covariance matrix (for stability)}

\item{r.truncation.p2}{= 5,  numeric used to limit extreme values in the covariance matrix (for stability)}

\item{burnin}{= 30  integer for the number of burn-in steps before sampling}

\item{thining}{= 10,  integer for the number of thining steps between sampling}

\item{length.p1}{= 100,  number of samples in phase 1}

\item{min.iter.p2}{= NULL,  minimum number of sub-steps in phase 2}

\item{max.iter.p2}{= NULL,  maximum number of sub-steps in phase 2}

\item{multiplication.iter.p2}{= 100, value for the lengths of sub-steps in phase 2 (multiplied by  2.52^k)}

\item{num.steps.p2}{= 6,  number of optimisation steps in phase 2}

\item{length.p3}{= 1000, number of samples in phase 3}

\item{neighborhood}{= c(0.7,0.3,0),  way of choosing partitions: probability vector (actors swap, merge/division, single actor move)}

\item{fixed.estimates}{= NULL, if some parameters are fixed, list with as many elements as effects, these elements equal a fixed value if needed, or NULL if they should be estimated}

\item{sizes.allowed}{= NULL,  vector of group sizes allowed in sampling (now, it only works for vectors like size_min:size_max)}

\item{sizes.simulated}{= NULL,  vector of group sizes allowed in the Markov chain but not necessraily sampled (now, it only works for vectors like size_min:size_max)}

\item{double.averaging}{= F,  option to average the statistics sampled in each sub-step of phase 2}

\item{inv.zcov}{= NULL,  initial value of the inverted covariance matrix (if a phase 3 was run before) to bypass the phase 1}

\item{inv.scaling}{= NULL,  initial value of the inverted scaling matrix (if a phase 3 was run before) to bypass the phase 1}

\item{parallel}{= F,  whether the phase 1 and 3 should be parallelized}

\item{parallel2}{= F,  whether there should be several phases 2 run in parallel}

\item{cpus}{= 1 ,how many cores can be used}
}
\value{
A list estimates 3 phases
}
\description{
Estimate ERPM
}
