% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_estimate.R
\name{estimate_ERPM_p3}
\alias{estimate_ERPM_p3}
\title{Estimate ERPM phase 3}
\usage{
estimate_ERPM_p3(
  partition,
  nodes,
  objects,
  effects,
  startingestimates,
  burnin = 30,
  thining = 10,
  length.p3 = 1000,
  neighborhood = c(0.7, 0.3, 0),
  fixed.estimates = NULL,
  sizes.allowed = NULL,
  sizes.simulated = NULL
)
}
\arguments{
\item{partition}{observed partition}

\item{nodes}{nodeset (data frame)}

\item{objects}{objects used for statistics calculation (list with a vector "name", and a vector "object")}

\item{effects}{effects/sufficient statistics (list with a vector "names", and a vector "objects")}

\item{startingestimates}{first guess for the model parameters}

\item{burnin}{= 30  integer for the number of burn-in steps before sampling}

\item{thining}{= 10,  integer for the number of thining steps between sampling}

\item{length.p3}{= 1000, number of samples in phase 3}

\item{neighborhood}{= c(0.7,0.3,0),  way of choosing partitions: probability vector (actors swap, merge/division, single actor move)}

\item{fixed.estimates}{= NULL, if some parameters are fixed, list with as many elements as effects, these elements equal a fixed value if needed, or NULL if they should be estimated}

\item{sizes.allowed}{= NULL,  vector of group sizes allowed in sampling (now, it only works for vectors like size_min:size_max)}

\item{sizes.simulated}{= NULL,  vector of group sizes allowed in the Markov chain but not necessraily sampled (now, it only works for vectors like size_min:size_max)}

\item{multiplicationfactor}{= 30,  for now, useless}
}
\value{
A estimates phase 3
}
\description{
Estimate ERPM phase 3
}
