% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_loglikelihood.R
\name{estimate_logL}
\alias{estimate_logL}
\title{Estimate log likelihood}
\usage{
estimate_logL(
  partition,
  nodes,
  effects,
  objects,
  theta,
  theta_0,
  M,
  num.steps,
  burnin,
  thining,
  neighborhoods = c(0.7, 0.3, 0),
  sizes.allowed = NULL,
  sizes.simulated = NULL,
  logL_0 = NULL,
  parallel = F,
  cpus = 1
)
}
\arguments{
\item{partition}{observed partition}

\item{nodes}{node set (data frame)}

\item{effects}{effects/sufficient statistics (list with a vector "names", and a vector "objects")}

\item{objects}{objects used for statistics calculation (list with a vector "name", and a vector "object")}

\item{theta}{estimated model parameters}

\item{theta_0}{model parameters if all other effects than "num-groups" are fixed to 0 (basic Dirichlet partition model)}

\item{M}{number of steps in the path-sampling algorithm}

\item{num.steps}{number of samples in each step}

\item{burnin}{integer for the number of burn-in steps before sampling}

\item{thining}{integer for the number of thining steps between sampling}

\item{neighborhoods}{= c(0.7,0.3,0) way of choosing partitions}

\item{sizes.allowed}{= NULL,   vector of group sizes allowed in sampling (now, it only works for vectors like size_min:size_max)}

\item{sizes.simulated}{= NULL,  vector of group sizes allowed in the Markov chain but not necessraily sampled (now, it only works for vectors like size_min:size_max)}

\item{logL_0}{= NULL, if known, the value of the log likelihood of the basic dirichlet model}
}
\value{
List with the log likelihood , AIC, lambda and the draws
}
\description{
Estimate log likelihood
}
