% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_estimate.R
\name{estimate_multipleBERPM}
\alias{estimate_multipleBERPM}
\title{Estimation ERPM for multiple observations}
\usage{
estimate_multipleBERPM(
  partitions,
  presence.tables,
  nodes,
  objects,
  effects,
  mean.priors,
  sd.priors,
  start.chains = NULL,
  burnin.1 = 30,
  thining.1 = 10,
  num.chains = 3,
  length.chains = 1000,
  burnin.2 = 30,
  neighborhood.partition = c(0.7, 0.3, 0, 0, 0, 0),
  neighborhood.augmentation = NULL,
  sizes.allowed = NULL,
  sizes.simulated = NULL,
  parallel = F,
  cpus = 1
)
}
\arguments{
\item{partitions}{observed partitions}

\item{presence.tables}{matrix indicating which actors were present for each observations (mandatory)}

\item{nodes}{nodeset (data frame)}

\item{objects}{objects used for statistics calculation (list with a vector "name", and a vector "object")}

\item{effects}{effects/sufficient statistics (list with a vector "names", and a vector "objects")}

\item{mean.priors, }{means of the normal distributions of prior parameters}

\item{sd.priors, }{standard deviations of the normal distributions of prior parameters}

\item{start.chains}{= NULL,  define a list of starting values for parameters}

\item{burnin.1}{= 30, integer for the number of burn-in steps before sampling in the main MCMC chain}

\item{thining.1}{= 10, integer for the number of thining steps between sampling in the main MCMC chain}

\item{num.chains}{= 3, number of MChains}

\item{length.chains}{= 1000, number of samples of each chain}

\item{burnin.2}{= 30, integer for the number of burn-in steps before sampling int the MCMC to sample partitions}

\item{neighborhood.partition}{= c(0.7,0.3,0,0,0,0), way of choosing partitions: probability vector (actors swap, merge/division, single actor move, pair move)}

\item{neighborhood.augmentation}{= NULL, standard deviations auround the parameters to draw the augmented distrubtion}

\item{sizes.allowed}{= NULL, vector of group sizes allowed in sampling (now, it only works for vectors like size_min:size_max)}

\item{sizes.simulated}{= NULL, vector of group sizes allowed in the Markov chain but not necessraily sampled (now, it only works for vectors like size_min:size_max)}

\item{parallel}{= F, whether the calculation of stats should be parallelized}

\item{cpus}{= 1 , how many cores can be used}
}
\value{
A list of estimates
}
\description{
Estimation ERPM for multiple observations
}
