% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exactcalculations.R
\name{exactestimates_numgroups}
\alias{exactestimates_numgroups}
\title{Exact estimates number of groups}
\usage{
exactestimates_numgroups(num.nodes, pmin, pmax, pinc)
}
\arguments{
\item{num.nodes}{XXX}

\item{pmin}{XXX}

\item{pmax}{XXX}
}
\value{
XXX
}
\description{
Exact estimates number of groups
}
