% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{find_all_partitions}
\alias{find_all_partitions}
\title{Function to enumerate all partitions for a given n}
\usage{
find_all_partitions(n)
}
\arguments{
\item{n}{XXX}
}
\value{
XXX
}
\description{
Function to enumerate all partitions for a given n
}
