% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_burninthining.R
\name{gridsearch_burnin_single}
\alias{gridsearch_burnin_single}
\title{Grid - search burning single}
\usage{
gridsearch_burnin_single(
  partition,
  theta,
  nodes,
  effects,
  objects,
  num.steps,
  neighborhoods,
  sizes.allowed,
  sizes.simulated,
  parallel = F,
  cpus = 1
)
}
\arguments{
\item{partition}{A partition (vector)}

\item{theta}{Initial model parameters}

\item{nodes}{Node set (data frame)}

\item{effects}{Effects/sufficient statistics (list with a vector "names", and a vector "objects")}

\item{objects}{Objects used for statistics calculation (list with a vector "name", and a vector "object")}

\item{num.steps}{Number of samples wanted}

\item{neighborhoods}{List of probability vectors (proba actors swap, proba merge/division, proba single actor move)}

\item{sizes.allowed}{Vector of group sizes allowed in sampling (now, it only works for vectors like size_min:size_max)}

\item{sizes.simulated}{Vector of group sizes allowed in the Markov chain but not necessraily sampled (now, it only works for vectors like size_min:size_max)}

\item{parallel}{False, to run different neighborhoods in parallel}

\item{cpus}{Equal to 1}
}
\value{
all simulations
}
\description{
Grid - search burning single
}
