% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_statistics.R
\name{group_size}
\alias{group_size}
\title{Statistics on the size of groups in a partition}
\usage{
group_size(partition, stat)
}
\arguments{
\item{partition}{A partition (vector)}

\item{stat}{The statistic to compute : 'avg' for average and 'sd' for standard deviation}
}
\value{
A number corresponding to the correlation coefficient if the attribute is numerical or
the correlation ratio if the attribute is categorical.
}
\description{
This function computes the average or the standard
deviation of the size of groups in a partition.
}
\examples{
p <- c(1, 2, 2, 3, 3, 4, 4, 4, 5)
group_size(p,'avg')
group_size(p,'sd')
}
