% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmProposal.ErpmSplitStep.R
\name{InitErgmProposal.ErpmSplitStep}
\alias{InitErgmProposal.ErpmSplitStep}
\alias{ErpmSplitStep}
\title{ERGM proposal: ErpmSplitStep}
\usage{
InitErgmProposal.ErpmSplitStep(nw, ...)
}
\value{
A list describing the proposal for ergm.
}
\description{
ErpmSplitStep is a Metropolis-Hastings proposal used in ERPM when a partition
is represented as a bipartite actor-group membership network under the
constraint ~ b1part.

The move selects a non-empty group g, picks a subset of its actors, selects an
empty group vertex g', and moves the selected subset from g to g'. In the
bipartite encoding, this corresponds to a multi-toggle move of size 2*|S|:
for each actor i in S, remove (i, g) and add (i, g').
}
\details{
This function only registers the proposal with ergm. The actual move logic
is implemented in C in MH_ErpmSplitStep.
}
\author{
Jérémie Chichignoud - Cub'itech
}
\keyword{ERGM}
\keyword{ERPM}
\keyword{MCMC}
\keyword{b1part}
\keyword{proposal}
\keyword{split}
