% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.groups.R
\name{InitErgmTerm.groups}
\alias{InitErgmTerm.groups}
\title{ERGM term: groups
Compte les sommets dont le degré total (in+out) ∈ [from,to).
Delta-stat = somme des variations sur tail et head lors d'un toggle.
Arguments:
from (num), to (num, Inf autorisé).
Retour:
list(name="groups", inputs= c(rbind(from,to)), dependence=TRUE, emptynwstats=…)}
\usage{
InitErgmTerm.groups(nw, arglist, ..., version = packageVersion("ergm"))
}
\description{
ERGM term: groups
Compte les sommets dont le degré total (in+out) ∈ [from,to).
Delta-stat = somme des variations sur tail et head lors d'un toggle.
Arguments:
from (num), to (num, Inf autorisé).
Retour:
list(name="groups", inputs= c(rbind(from,to)), dependence=TRUE, emptynwstats=…)
}
