% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{order_groupids}
\alias{order_groupids}
\title{Function to replace the ids of the group without forgetting an id
and put in the first appearance order
for example: \verb{[2 1 1 4 2]} becomes \verb{[1 2 2 3 1]}}
\usage{
order_groupids(partition)
}
\arguments{
\item{partition}{observed partition}
}
\value{
a vector (partition)
}
\description{
Function to replace the ids of the group without forgetting an id
and put in the first appearance order
for example: \verb{[2 1 1 4 2]} becomes \verb{[1 2 2 3 1]}
}
