% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomeObjects.R
\docType{data}
\name{outcomeObjects}
\alias{outcomeObjects}
\alias{estimation}
\title{Exemplary outcome objects for the ERPM Package}
\format{
\code{estimation}
An results object created by the function \code{\link[=estimate_ERPM]{estimate_ERPM()}}.
}
\description{
These are exemplary outcome objects for the ERPM package and can be used in
order not to run all precedent functions and thus save time. The following
products are provided:
}
\keyword{datasets}
