% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phase1.R
\name{phase1}
\alias{phase1}
\title{Core function for Phase 1}
\usage{
phase1(
  startingestimates,
  inv.zcov,
  inv.scaling,
  z.phase1,
  z.obs,
  nodes,
  effects,
  objects,
  r.truncation.p1,
  length.p1,
  fixed.estimates
)
}
\arguments{
\item{startingestimates}{XXX}

\item{inv.zcov}{XXX}

\item{inv.scaling}{XXX}

\item{z.phase1}{XXX}

\item{z.obs}{XXX}

\item{nodes}{node set (data frame)}

\item{effects}{effects/sufficient statistics (list with a vector "names", and a vector "objects")}

\item{objects}{objects used for statistics calculation (list with a vector "name", and a vector "object")}

\item{r.truncation.p1}{XXX}

\item{length.p1}{XXX}

\item{fixed.estimates}{XXX}
}
\value{
XXX
}
\description{
Core function for Phase 1
}
