% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phase3.R
\name{phase3}
\alias{phase3}
\title{Core function for Phase 3}
\usage{
phase3(
  estimates.phase2,
  z.phase3,
  z.obs,
  nodes,
  effects,
  length.p3,
  fixed.estimates
)
}
\arguments{
\item{estimates.phase2}{XXX}

\item{z.phase3}{XXX}

\item{z.obs}{XXX}

\item{nodes}{XXX}

\item{effects}{XXX}

\item{length.p3}{XXX}

\item{fixed.estimates}{XXX}
}
\value{
XXX
}
\description{
Core function for Phase 3
}
