% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exactcalculations.R
\name{plot_averagesizes}
\alias{plot_averagesizes}
\title{Plot the average size distribution}
\usage{
plot_averagesizes(nmin, nmax, ninc)
}
\arguments{
\item{nmin}{A partition (vector)}
}
\description{
Plot the average size distribution
}
