% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exactcalculations.R
\name{plot_numgroups_likelihood}
\alias{plot_numgroups_likelihood}
\title{Plot likelihood of number groups}
\usage{
plot_numgroups_likelihood(m.obs, num.nodes, pmin, pmax, pinc)
}
\arguments{
\item{m.obs}{observed number of groups}

\item{num.nodes}{number of nodes}

\item{pmin}{lowest parameter value}

\item{pmax}{highest parameter value}

\item{pinc}{increment between different parameter values}
}
\value{
a vector
}
\description{
Function to plot the log-likelihood of the model with a single statistic (number of groups)
depending on the parameter value for this statistic
}
