% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_visualisation.R
\name{plot_partition}
\alias{plot_partition}
\title{Visualization of partition}
\usage{
plot_partition(
  partition,
  title = NULL,
  group.color = NULL,
  attribute.color = NULL,
  attribute.shape = NULL,
  palette.group = NULL,
  palette.attribute = NULL
)
}
\arguments{
\item{partition}{A partition (vector)}

\item{title}{Character, the title of the plot (default=NULL)}

\item{group.color}{A vector with the colors of the groups (default=NULL)}

\item{attribute.color}{A vector, attribute to represent with colors (default=NULL)}

\item{attribute.shape}{A vector, attribute to represent with shapes (default=NULL)}
}
\value{
A plot of the partition
}
\description{
This function plot the groups of a partition
}
\examples{
p <- c(1,1,1,2,2,2,2,3,3,3,4,4,4,4,4,4)
attr1 <- c(1,0,0,1,0,0,1,0,1,0,1,1,1,1,1,2)
attr2 <- c(1,1,1,1,0,0,3,0,1,0,1,1,1,1,1,2)
plot_partition(p,attribute.color = attr1, attribute.shape = attr2)
}
