% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_output.R
\name{print.results.bayesian.erpm}
\alias{print.results.bayesian.erpm}
\title{Print results of bayesian estimation (beta version)}
\usage{
\method{print}{results.bayesian.erpm}(result)
}
\description{
Print results of bayesian estimation (beta version)
}
