% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_output.R
\name{print.results.list.erpm}
\alias{print.results.list.erpm}
\title{Print estimation results}
\usage{
\method{print}{results.list.erpm}(x, ...)
}
\arguments{
\item{x}{output of the estimate function}

\item{...}{For internal use only.}
}
\value{
a data frame
}
\description{
Print estimation results
}
