% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_output.R
\name{print_results}
\alias{print_results}
\title{Print estimation results}
\usage{
print_results(result)
}
\arguments{
\item{result}{output of the estimate function}
}
\value{
a data frame
}
\description{
Print estimation results
}
