% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_utility.R
\name{proportion_isolate}
\alias{proportion_isolate}
\title{Proportion of isolates}
\usage{
proportion_isolate(partition)
}
\arguments{
\item{partition}{A partition (vector)}
}
\value{
A number corresponding to proportion of individuals alone.
}
\description{
This function computes the proportion of individuals not joining others.
}
\examples{
p <- c(1,2,2,3,3,4,4,4,5)
proportion_isolate(p)
}
