% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Metropolis.R
\name{reachable}
\alias{reachable}
\title{Reachable}
\usage{
reachable(i, partition1, partition2)
}
\arguments{
\item{i}{XXX}

\item{partition1}{XXX}

\item{partition2}{XXX}
}
\value{
A boolean
}
\description{
Reachable
}
