% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Metropolis.R
\name{reachable_p1}
\alias{reachable_p1}
\title{reachable neighborhood 1}
\usage{
reachable_p1(partition1, partition2)
}
\arguments{
\item{partition1}{XXX}

\item{partition2}{XXX}
}
\value{
A boolean
}
\description{
reachable neighborhood 1
}
