% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phase1.R
\name{run_phase1_multiple}
\alias{run_phase1_multiple}
\title{Phase 1 wrapper for multiple observations}
\usage{
run_phase1_multiple(
  partitions,
  startingestimates,
  z.obs,
  presence.tables,
  nodes,
  effects,
  objects,
  burnin,
  thining,
  gainfactor,
  a.scaling,
  r.truncation.p1,
  length.p1,
  neighborhood,
  fixed.estimates,
  sizes.allowed,
  sizes.simulated,
  parallel = F,
  cpus = 1
)
}
\arguments{
\item{partitions}{observed partitions}

\item{startingestimates}{XXX}

\item{z.obs}{XXX}

\item{presence.tables}{XXX}

\item{nodes}{node set (data frame)}

\item{effects}{effects/sufficient statistics (list with a vector "names", and a vector "objects")}

\item{objects}{objects used for statistics calculation (list with a vector "name", and a vector "object")}

\item{burnin}{integer for the number of burn-in steps before sampling}

\item{thining}{integer for the number of thining steps between sampling}

\item{gainfactor}{XXX}

\item{a.scaling}{XXX}

\item{r.truncation.p1}{XXX}

\item{length.p1}{XXX}

\item{neighborhood}{= c(0.7,0.3,0) way of choosing partitions}

\item{fixed.estimates}{XXX}

\item{sizes.allowed}{= NULL,   vector of group sizes allowed in sampling (now, it only works for vectors like size_min:size_max)}

\item{sizes.simulated}{= NULL,  vector of group sizes allowed in the Markov chain but not necessraily sampled (now, it only works for vectors like size_min:size_max)}

\item{parallel}{XXX}

\item{cpus}{= 1}
}
\value{
XXX
}
\description{
Phase 1 wrapper for multiple observations
}
