% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phase3.R
\name{run_phase3_single}
\alias{run_phase3_single}
\title{Phase 3 wrapper for single observation}
\usage{
run_phase3_single(
  partition,
  estimates.phase2,
  z.obs,
  nodes,
  effects,
  objects,
  burnin,
  thining,
  a.scaling,
  length.p3,
  neighborhood,
  sizes.allowed,
  sizes.simulated,
  fixed.estimates,
  parallel = F,
  cpus = 1
)
}
\arguments{
\item{partition}{XXX}

\item{estimates.phase2}{XXX}

\item{z.obs}{XXX}

\item{nodes}{XXX}

\item{effects}{XXX}

\item{objects}{XXX}

\item{burnin}{XXX}

\item{thining}{XXX}

\item{a.scaling}{XXX}

\item{length.p3}{XXX}

\item{neighborhood}{XXX}

\item{sizes.allowed}{XXX}

\item{sizes.simulated}{XXX}

\item{fixed.estimates}{XXX}

\item{parallel}{XXX}

\item{cpus}{XXX}
}
\value{
XXX
}
\description{
Phase 3 wrapper for single observation
}
