% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Metropolis.R
\name{sample_new_partition_p3_restricted}
\alias{sample_new_partition_p3_restricted}
\title{sample partition neighborhood 3 restricted}
\usage{
sample_new_partition_p3_restricted(
  current.partition,
  size_neighborhood,
  sizes.simulated
)
}
\arguments{
\item{current.partition}{XXX}

\item{size_neighborhood}{XXX}

\item{sizes.simulated}{XXX}
}
\value{
A list
}
\description{
sample partition neighborhood 3 restricted
}
