% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Metropolis.R
\name{sample_new_partition_p4}
\alias{sample_new_partition_p4}
\title{sample new partition phase 4}
\usage{
sample_new_partition_p4(current.partition, size_neighborhood)
}
\value{
A partition
}
\description{
sample new partition phase 4
}
