% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Metropolis.R
\name{sample_new_partition_p6}
\alias{sample_new_partition_p6}
\title{sample new partition phase 6}
\usage{
sample_new_partition_p6(current.partition, size_neighborhood)
}
\value{
A partition
}
\description{
sample new partition phase 6
}
