% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Metropolis.R
\name{sample_new_partition_p7}
\alias{sample_new_partition_p7}
\title{sample new partition phase 7}
\usage{
sample_new_partition_p7(
  current.partition,
  current.partition2,
  size_neighborhood
)
}
\value{
A partition
}
\description{
sample new partition phase 7
}
