% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_utility.R
\name{stat_size}
\alias{stat_size}
\title{Statistics on the size of groups in a partition}
\usage{
stat_size(partition, stat)
}
\arguments{
\item{partition}{A partition (vector)}

\item{stat}{The statistic to compute : 'avg' for average and 'sd' for standard deviation}
}
\value{
A number corresponding to the correlation coefficient if the attribute is numerical or
the correlation ratio if the attribute is categorical.
}
\description{
This function computes the average or the standard
deviation of the size of groups in a partition.
}
\examples{
p <- c(1,2,2,3,3,4,4,4,5)
correlation_with_size(p,'avg')
correlation_with_size(p,'sd')
}
