% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{Stirling2_constraints}
\alias{Stirling2_constraints}
\title{Function to calculate the number of partitions with k groups
of sizes between smin and smax}
\usage{
Stirling2_constraints(n, k, smin, smax)
}
\arguments{
\item{n}{XXX}

\item{k}{XXX}

\item{smin}{XXX}

\item{smax}{XXX}
}
\value{
XXX
}
\description{
Function to calculate the number of partitions with k groups
of sizes between smin and smax
}
