% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{Bell_constraints}
\alias{Bell_constraints}
\title{Function to calculate the number of partitions with groups
of sizes between smin and smax}
\usage{
Bell_constraints(n, smin, smax)
}
\arguments{
\item{n}{XXX}

\item{smin}{XXX}

\item{smax}{XXX}
}
\value{
XXX
}
\description{
Function to calculate the number of partitions with groups
of sizes between smin and smax
}
