% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exactcalculations.R
\name{calculate_denominator_Dirichlet_restricted}
\alias{calculate_denominator_Dirichlet_restricted}
\title{Calculate Dirichlet denominator}
\usage{
calculate_denominator_Dirichlet_restricted(n, smin, smax, alpha, results)
}
\arguments{
\item{n}{number of nodes}

\item{smin}{minimum size for a group}

\item{smax}{maximum size for a group}

\item{alpha}{parameter value}

\item{results}{a list}
}
\value{
a numeric
}
\description{
Recursive function to calculate the denominator for the model with a single statistic
for the number of groups and a given parameter value.
The set of possible partitions can be restricted to partitions with groups of a certain size.
}
