% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exactcalculations.R
\name{calculate_proba_Dirichlet_restricted}
\alias{calculate_proba_Dirichlet_restricted}
\title{Plot likelihood of number groups}
\usage{
calculate_proba_Dirichlet_restricted(alpha, stat, n, smin, smax)
}
\description{
Plot likelihood of number groups
}
