% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exactcalculations.R
\name{calculate_proba_Dirichlet_restricted}
\alias{calculate_proba_Dirichlet_restricted}
\title{Estimation basic model with size constraint}
\usage{
calculate_proba_Dirichlet_restricted(alpha, stat, n, smin, smax)
}
\arguments{
\item{alpha}{XXX}

\item{stat}{XXX}

\item{n}{XXX}

\item{smin}{XXX}

\item{smax}{XXX}
}
\value{
XXX
}
\description{
Estimation basic model with size constraint
}
