% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{check_sizes}
\alias{check_sizes}
\title{Function to determine whether a partition contains the allowed group sizes}
\usage{
check_sizes(partition, sizes.allowed, numgroups.allowed)
}
\arguments{
\item{partition}{observed partition}

\item{sizes.allowed}{vector containing possible group sizes in the partition}

\item{numgroups.allowed}{vector containing possible number of groups in the partition}
}
\value{
boolean
}
\description{
Function to determine whether a partition contains the allowed group sizes
}
