% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exactcalculations.R
\name{compute_numgroups_denominator}
\alias{compute_numgroups_denominator}
\title{Recursive function to compute the value of the denominator}
\usage{
compute_numgroups_denominator(num.nodes, alpha)
}
\arguments{
\item{num.nodes}{XXX}

\item{alpha}{Node set (data frame)}
}
\value{
XXX
}
\description{
Recursive function to compute the value of the denominator
}
