% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Metropolis.R
\name{compute_size_neighborhood}
\alias{compute_size_neighborhood}
\title{compute size neighborhood}
\usage{
compute_size_neighborhood(
  i,
  partition,
  partition2 = NULL,
  sizes.simulated = NULL
)
}
\arguments{
\item{i}{XXX}

\item{partition2}{XXX}

\item{sizes.simulated}{XXX}

\item{partitions}{XXX}
}
\value{
A list
}
\description{
compute size neighborhood
}
