% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{count_classes}
\alias{count_classes}
\title{Function to count the number of partitions with a certain
group size structure, for all possible group size structure}
\usage{
count_classes(allpartitions)
}
\arguments{
\item{allpartitions}{XXX}
}
\value{
XXX
}
\description{
Function to count the number of partitions with a certain
group size structure, for all possible group size structure
}
