% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erpm-package.R
\docType{package}
\name{erpm-package}
\alias{erpm-package}
\alias{erpm}
\title{\code{erpm} package}
\description{
The erpm package is an R package that allows to fit an exponential partition model
to cross-sectional or longitudinal partition data.
}
\author{
\strong{Maintainer}: Marion Hoffman \email{marion.hoffman.31@gmail.com} (\href{https://orcid.org/0000-0002-0741-7760}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Alexandra Amani
}

}
