% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erpm-package.R
\docType{package}
\name{erpm-package}
\alias{erpm-package}
\alias{erpm}
\title{`erpm` package}
\description{
The erpm package is an R package that allows to fit an exponential partition model
to cross-sectional or longitudinal partition data.
}
