% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erpm-package.R
\docType{package}
\name{ERPM-package}
\alias{ERPM}
\alias{ERPM-package}
\title{ERPM: Exponential Random Partition Models}
\description{
Simulates and estimates the Exponential Random Partition Model presented in the paper Hoffman, Block, and Snijders (2023) \doi{10.1177/00811750221145166}. It can also be used to estimate longitudinal partitions, following the model proposed in Hoffman and Chabot (2023) \doi{10.1016/j.socnet.2023.04.002}. The model is an exponential family distribution on the space of partitions (sets of non-overlapping groups) and is called in reference to the Exponential Random Graph Models (ERGM) for networks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stocnet/ERPM}
  \item Report bugs at \url{https://github.com/stocnet/ERPM/issues}
}

}
\author{
\strong{Maintainer}: Marion Hoffman \email{marion.hoffman.31@gmail.com} (\href{https://orcid.org/0000-0002-0741-7760}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Alexandra Amani
  \item Nico Keiser
}

}
\keyword{internal}
