% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exactcalculations.R
\name{exactestimates_numgroups}
\alias{exactestimates_numgroups}
\title{Exact estimates number of groups}
\usage{
exactestimates_numgroups(num.nodes, pmin, pmax, pinc)
}
\arguments{
\item{num.nodes}{number of nodes}

\item{pmin}{lowest parameter value}

\item{pmax}{highest parameter value}

\item{pinc}{increment between different parameter values}
}
\value{
a list
}
\description{
This function finds the best estimate for a model only including the statistics of number of groups.
It does a grid search for a vector of potential parameters, for all numbers of groups.
}
