% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{find_all_partitions}
\alias{find_all_partitions}
\title{Function to enumerate all possible partitions for a given n}
\usage{
find_all_partitions(n)
}
\arguments{
\item{n}{number of nodes}
}
\value{
matrix where each line corresponds to a possible partition
}
\description{
Function to enumerate all possible partitions for a given n
}
\examples{
n <- 6
all_partitions <- find_all_partitions(n)

}
