% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmTerm.cliques.R
\name{InitErgmTerm.cliques}
\alias{InitErgmTerm.cliques}
\title{ERGM term: cliques
Terme composite simple: Δ = w * ( ΔTriangles + ΔTwoStars ) autour du toggle.
Arguments:
w (num) poids global, défaut 1.}
\usage{
InitErgmTerm.cliques(nw, arglist, ..., version = packageVersion("ergm"))
}
\description{
ERGM term: cliques
Terme composite simple: Δ = w * ( ΔTriangles + ΔTwoStars ) autour du toggle.
Arguments:
w (num) poids global, défaut 1.
}
