% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_utility.R
\name{number_ties}
\alias{number_ties}
\title{Same pairs of individuals in a partition}
\usage{
number_ties(partition, dyadic_attribute)
}
\arguments{
\item{partition}{A partition (vector)}

\item{attribute}{A vector containing the values of the attribute}

\item{stat}{The statistic to compute : 'avg_pergroup' for the average per group , 'sum_pergroup' for the sum, 'sum_perind' and 'avg_perind' for the number of ties per individuals
each individual has in its group.}
}
\value{
The statisic chosen in stat
}
\description{
This function computes the number of ties.
}
\examples{
p <- c(1,2,2,3,3,4,4,4,5)
at <- c(0,1,1,1,1,0,0,0,0)
number_ties(p,at,'avg')
}
