% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exactcalculations.R
\name{plot_averagesizes}
\alias{plot_averagesizes}
\title{Plot average sizes}
\usage{
plot_averagesizes(nmin, nmax, ninc)
}
\arguments{
\item{nmin}{minimum number of nodes}

\item{nmax}{maximum number of nodes}

\item{ninc}{increment between the different number of nodes}
}
\value{
a vector
}
\description{
Function to plot the average size of a random partition depending on the number of nodes
}
