% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exactcalculations.R
\name{plot_numgroups_likelihood}
\alias{plot_numgroups_likelihood}
\title{Plot likelihood of number groups}
\usage{
plot_numgroups_likelihood(m.obs, num.nodes, pmin, pmax, pinc)
}
\arguments{
\item{num.nodes}{XXX}

\item{pmin}{Node set (data frame)}

\item{pmax}{Effects/sufficient statistics (list with a vector "names", and a vector "objects")}

\item{pinc}{Objects used for statistics calculation (list with a vector "name", and a vector "object")}

\item{m.nodes}{A partition (vector)}
}
\value{
XXX
}
\description{
Plot likelihood of number groups
}
