% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_output.R
\name{print.results.bayesian.erpm}
\alias{print.results.bayesian.erpm}
\title{Print results of bayesian estimation (beta version)}
\usage{
\method{print}{results.bayesian.erpm}(x, ...)
}
\arguments{
\item{x}{output of the bayesian estimate function}

\item{...}{For internal use only.}
}
\value{
a data frame
}
\description{
Print results of bayesian estimation (beta version)
}
