% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_output.R
\name{print.results.p3.erpm}
\alias{print.results.p3.erpm}
\title{Print results of estimation of phase 3}
\usage{
\method{print}{results.p3.erpm}(x, ...)
}
\arguments{
\item{x}{output of the estimate function}

\item{...}{For internal use only.}
}
\value{
a data frame
}
\description{
Print results of estimation of phase 3
}
