% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phase2.R
\name{run_phase2_single}
\alias{run_phase2_single}
\title{Run phase 2 single}
\usage{
run_phase2_single(
  partition,
  estimates.phase1,
  inv.zcov,
  inv.scaling,
  z.obs,
  nodes,
  effects,
  objects,
  burnin,
  thining,
  num.steps,
  gainfactors,
  r.truncation.p2,
  min.iter,
  max.iter,
  multiplication.iter,
  neighborhood,
  fixed.estimates,
  sizes.allowed,
  sizes.simulated,
  double.averaging
)
}
\description{
Run phase 2 single
}
