% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Metropolis.R
\name{sample_new_partition}
\alias{sample_new_partition}
\title{sample new partition}
\usage{
sample_new_partition(
  i,
  current.partition,
  size_neighborhood,
  numgroups.simulated = NULL,
  sizes.simulated = NULL
)
}
\arguments{
\item{i}{XXX}

\item{current.partition}{XXX}

\item{size_neighborhood}{XXX}

\item{numgroups.simulated}{XXX}

\item{sizes.simulated}{XXX}
}
\value{
A partition
}
\description{
sample new partition
}
