% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_burninthining.R
\name{simulate_thining_single}
\alias{simulate_thining_single}
\title{Simulate thining single}
\usage{
simulate_thining_single(
  partition,
  theta,
  nodes,
  effects,
  objects,
  num.steps,
  neighborhood,
  sizes.allowed,
  sizes.simulated,
  burnin,
  max.thining
)
}
\arguments{
\item{partition}{A partition (vector)}

\item{theta}{Initial model parameters}

\item{nodes}{Node set (data frame)}

\item{effects}{Effects/sufficient statistics (list with a vector "names", and a vector "objects")}

\item{objects}{Objects used for statistics calculation (list with a vector "name", and a vector "object")}

\item{num.steps}{Number of samples wanted}

\item{neighborhood}{Way of choosing partitions: probability vector (proba actors swap, proba merge/division, proba single actor move)}

\item{sizes.allowed}{Vector of group sizes allowed in sampling (now, it only works for vectors like size_min:size_max)}

\item{sizes.simulated}{Vector of group sizes allowed in the Markov chain but not necessraily sampled (now, it only works for vectors like size_min:size_max)}

\item{burnin}{XXX}

\item{max.thining}{XXX}
}
\value{
A list
}
\description{
Simulate thining single
}
