% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{Stirling2_constraints}
\alias{Stirling2_constraints}
\title{Function to calculate the number of partitions with k groups
of sizes between smin and smax}
\usage{
Stirling2_constraints(n, k, smin, smax)
}
\arguments{
\item{n}{number of nodes}

\item{k}{number of groups}

\item{smin}{minimum group size possible in the partition}

\item{smax}{maximum group size possible in the partition}
}
\value{
a numeric
}
\description{
Function to calculate the number of partitions with k groups
of sizes between smin and smax
}
\examples{
n <- 6
k <- 2
size_min <- 2
size_max <- 4
Stirling2_constraints(n, k, size_min, size_max)

}
